requirements.txt
# local package
-e .

# external requirements
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.4.0
transformers~=4.16.2
torch~=1.10.2
src~=0.1.0
nltk~=3.6.7
bs4~=0.0.1
beautifulsoup4~=4.10.0
numpy~=1.22.1
wandb~=0.12.10
sklearn~=0.0
scikit-learn~=1.0.2
omegaconf~=2.1.1
setuptools~=59.5.0
hydra-core~= 1.1.1
